Enhanced retain tests and move into test/retain.js
