+++ test/qos1.js
@@ -383,109 +383,6 @@ test('do not resend QoS 1 packets at reconnect if puback was received', function
   })
 })
 
-test('deliver QoS 1 retained messages', function (t) {
-  var broker = aedes()
-  var publisher = connect(setup(broker))
-  var subscriber = connect(setup(broker))
-  var expected = {
-    cmd: 'publish',
-    topic: 'hello',
-    payload: Buffer.from('world'),
-    qos: 0,
-    dup: false,
-    length: 12,
-    retain: true
-  }
-
-  publisher.inStream.write({
-    cmd: 'publish',
-    topic: 'hello',
-    payload: 'world',
-    qos: 1,
-    messageId: 42,
-    retain: true
-  })
-
-  publisher.outStream.on('data', function (packet) {
-    subscribe(t, subscriber, 'hello', 1, function () {
-      subscriber.outStream.once('data', function (packet) {
-        t.deepEqual(packet, expected, 'packet must match')
-        t.end()
-      })
-    })
-  })
-})
-
-test('deliver QoS 1 retained messages', function (t) {
-  var broker = aedes()
-  var publisher = connect(setup(broker))
-  var subscriber = connect(setup(broker))
-  var expected = {
-    cmd: 'publish',
-    topic: 'hello',
-    payload: Buffer.from('world'),
-    qos: 1,
-    dup: false,
-    length: 14,
-    retain: false
-  }
-
-  subscribe(t, subscriber, 'hello', 1, function () {
-    subscriber.outStream.once('data', function (packet) {
-      delete packet.messageId
-      t.deepEqual(packet, expected, 'packet must match')
-      t.end()
-    })
-    publisher.inStream.write({
-      cmd: 'publish',
-      topic: 'hello',
-      payload: 'world',
-      qos: 1,
-      messageId: 42,
-      retain: true
-    })
-  })
-})
-
-test('deliver QoS 0 retained message with QoS 1 subscription', function (t) {
-  var broker = aedes()
-  var publisher = connect(setup(broker))
-  var subscriber = connect(setup(broker))
-  var expected = {
-    cmd: 'publish',
-    topic: 'hello',
-    payload: Buffer.from('world'),
-    qos: 0,
-    dup: false,
-    length: 12,
-    retain: true
-  }
-
-  broker.mq.on('hello', function (msg, cb) {
-    cb()
-
-    // defer this or it will receive the message which
-    // is being published
-    setImmediate(function () {
-      subscribe(t, subscriber, 'hello', 1, function () {
-        subscriber.outStream.once('data', function (packet) {
-          t.deepEqual(packet, expected, 'packet must match')
-          t.end()
-        })
-      })
-    })
-  })
-
-  publisher.inStream.write({
-    cmd: 'publish',
-    topic: 'hello',
-    payload: Buffer.from('world'),
-    qos: 0,
-    messageId: 42,
-    retain: true
-  })
-})
-
 test('remove stored subscriptions after unsubscribe', function (t) {
   var broker = aedes()
   var publisher
@@ -598,78 +495,6 @@ test('downgrade QoS 0 publish on QoS 1 subsciption', function (t) {
       topic: 'hello',
       payload: 'world',
       qos: 0
-    })
-  })
-})
-
-test('not clean and retain messages with QoS 1', function (t) {
-  var broker = aedes()
-  var publisher
-  var subscriber = connect(setup(broker), { clean: false, clientId: 'abcde' })
-  var expected = {
-    cmd: 'publish',
-    topic: 'hello',
-    payload: Buffer.from('world'),
-    qos: 1,
-    dup: false,
-    length: 14,
-    retain: true
-  }
-
-  subscribe(t, subscriber, 'hello', 1, function () {
-    subscriber.inStream.write({
-      cmd: 'disconnect'
-    })
-
-    subscriber.outStream.on('data', function (packet) {
-      console.log('original', packet)
-    })
-
-    publisher = connect(setup(broker))
-
-    publisher.inStream.write({
-      cmd: 'publish',
-      topic: 'hello',
-      payload: 'world',
-      qos: 1,
-      messageId: 42,
-      retain: true
-    })
-
-    publisher.outStream.once('data', function (packet) {
-      t.equal(packet.cmd, 'puback')
-
-      broker.on('clientError', function (client, err) {
-        t.fail('no error')
-      })
-
-      subscriber = connect(setup(broker), { clean: false, clientId: 'abcde' }, function (connect) {
-        t.equal(connect.sessionPresent, true, 'session present is set to true')
-      })
-
-      subscriber.outStream.once('data', function (packet) {
-        t.notEqual(packet.messageId, 42, 'messageId must differ')
-        t.equal(packet.qos, 0, 'qos degraded to 0 for retained')
-        var prevId = packet.messageId
-        delete packet.messageId
-        packet.qos = 1
-        packet.length = 14
-        t.deepEqual(packet, expected, 'packet must match')
-
-        // message is duplicated
-        subscriber.outStream.once('data', function (packet2) {
-          var curId = packet2.messageId
-          t.notOk(curId === prevId, 'messageId must differ')
-          subscriber.inStream.write({
-            cmd: 'puback',
-            messageId: curId
-          })
-          delete packet2.messageId
-          t.deepEqual(packet, expected, 'packet must match')
-
-          t.end()
-        })
-      })
     })
   })
 })
